add soldering note to CC1352P-2 section (#212)
To use the CC1352P-2's external SMA antenna connector, a very tiny capacitor must be resoldered. This should be noted as it is not that easy and some people might select the CC1352P-2 (over the CC26X2R1) just for that antenna connector (not knowing that this requires soldering skills / additional work).