Reorganise home page links
I have found the homepage links to be a bit difficult to navigate at first. The link to running Zigbee2mqtt on bare-metal linux was difficult to see at first because my eyes jumped straight to the bullet points. The other categories appear to be overlapping so I have tried to simplify this.

All good if you don't agree with the changes, this is just a suggestion. Feel free to decline the pull request or amend. Thank you