Define empty string fo input_text
Without initial definition the input fields are inactive in Lovelace (users are not able to put in text directly in the field on a card) and show "unknown". This is not really user friendly and maybe even confusing. Defining an empty string makes the input field active and initially shows "(empty value)" which makes more sense.