Added missing details about exposed properties
After updating my zigbee2mqtt bridge from 1.13.0 to 1.17.0 I noticed that this plug is now exposing 4 new details that are not referenced in the documentation.
New properties are Voltage, Energy, Power and Current.
Added to both the "Exposes" section and also to the "Manual Configuration". Copied a lot of the text from a similar Xiaomi plug as it's exactly the same properties and types.
As a note, and this is probably a comment for the herdsman project, the HA integration is automatically creating a power_outage_memory sensor that the plug doesn't seem to support.