FAQ.md: Mention a hacky method of "fixing" the non-responsive CC2531 dongle
Spent quite a while looking for a way to make my freshly flashed dongle work, tried everything on this page and elsewhere until finding a working solution in [this obscure comment](https://www.gitmemory.com/issue/Koenkk/zigbee2mqtt/2997/605655045). Plugging the dongle into a laptop and then back actually worked, and I'm apparently not the only one with this issue, so it this method is for sure worth mentioning here (it would be especially nice if someone could explain what is actually happening, but the advice is valuable even without this information).