Update calibration instructions for Tuya TS130F
According to the instructions of the product, you need to press the "pause" button after fully closed. I tried without it and it did not work, with it all worked perfectly.

I've removed too the mosquitto command to put it into calibration mode, I think it can produce confusion to people. There are different ways to enable/disable de calibration mode.

The LoraTap white-label image is quite different from the Tuya one. Is there a way to add different images for the same product here? Can I simply add another Picture to the table?