Update AC26940_AC31266.md
It seems that Ledvance uses the same identifier for multiple devices. I've got an Ledvance Smart+ Nightlight Plug which identifies as LEDVANCE AC26940/AC31266. I've updated the device file (not sure if that's the correct way though, due to the warning that this page was auto-generated (but I can't find the source)).

The referenced image should maybe also be changed.