Update BSEN-C2.md
I spent/waited quite some time (years!) to find a door sensor,Z2M compatible and suitable for outdoor use and finally found this one! 
This is great for whoever want to use such sensor on a garden gate, letter box or whatever outdoor use.
IP45 level is as well as "indoor/outdoor use" information is from Bosch documentation.
I'm not sure if it's against any rules or best pratice to mention technical characteristics on a sensor page, but I thought it might be interesting for other people that might seek outdoor door sensor too.
Feel free to cancel if not compliant.