Add section about install code
According to the manual https://www.bosch-smarthome.com/de/media/content/10_bedienungsanleitungen/bewegungsmelder/8-750-000-920_v001_shmd_qsg_de_20200211_web.pdf you need to also scan an install code with this device.
There is already an issue (https://github.com/Koenkk/zigbee2mqtt/issues/17315) because this is missing in the documentation. 
If the install code solves the issue problem, I think it should be added to the documentation.