Update documentation to include the pairing process via install code (#2019)
According to the manual https://www.bosch-smarthome.com/de/de/service/hilfe/hilfe-zum-produkt/hilfe-zum-wassermelder/ you need to also scan an install code with this device. There was already an issue (Koenkk/zigbee2mqtt#17315) on the motion sensor because this is missing in the documentation. I think all currently supported Bosch devices that need an Install Code should have this in their documentation now. 
@Koenkk is there any way to make sure this is added for new Bosch devices in the future (as they will probably keep this going).