Update README.md
If you change from zStack 1.4.0 to 3.0, as is my case, you need to re-pair your routers most likely. It is mentioned sometimes in the past, but it is not in the FAQ. For me, as a CC2531 user that finally decided to upgrade, I ran into the issue that none of my routers want to connect, only the end devices. Therefore, this breaks my network and I need to re-pair at least my routers, thereby hopefully automatically restoring access to all end nodes as well.