Workaround for "failed interview" in E1525_E1745.md
This describes a workaround for a "failed interview" trying to re-add a previously working sensor to Z2M after the device batteries expire and are replaced. It is possibly happening to devices that have device bindings. 

I found the work around in this github issue: https://github.com/Koenkk/zigbee2mqtt/issues/12422#issuecomment-2178391096