Add note for MQTT server hostname usage
When Z2M is used as a home assistant addon, I've noticed that using `localhost` won't work for to connect to Mosquito broker.

The connection is established, but later stopped with error `z2m: Connecting to MQTT server at mqtt://localhost:1883`. The broker does see the connection, just doesn't process it for some reason.

A change from `localhost` to `core-mosquito`, `mosquito` or `homeassistant` seems to solve the issue.