Update 067776.md
Discovered a bug in the device. Without this extra step that I just added calibration will succeed and you will be able to set its position, but with a catch. If you are setting a lower position from a higher one (e.g. from a 100 you set it to 90) then everything will work correctly. But if you try to go to a higher position (e.g. from 90 to 91, or 0 to 50) then it will always fully open. With the additional step that I added to the documentation this problem seems to be resolved.